1. http的request解析(bug)；

+++ hhou/parser/HH_Parse.h
@@ -33,7 +33,7 @@ namespace hhou
         virtual ~HHParse() {}
 
         /**先进行必要信息的解析(错误数据返回-1，数据不完整返回0，接收完全返回>0)*/
-        virtual int ParseData(const char *buf, size_t nLen, string &strRet);
+        virtual int ParseData(const char *buf, int nLen, string &strRet);
     };
 }
